requirements.txt
requests==2.32.4
websockets==14.1
python-dateutil==2.8.0
python-dotenv==1.0.0
poke-engine==0.0.46 --config-settings="build-args=--features poke-engine/terastallization --no-default-features"
